% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.data.prep.R
\name{fe.data.prep}
\alias{fe.data.prep}
\title{Prepares data for model fitting (fe.prov)}
\usage{
fe.data.prep(data, Y.char, Z.char, prov.char, cutoff = 10)
}
\arguments{
\item{data}{a \code{data.frame} including response, provider ID, and covariates, with missing values imputed}

\item{Y.char}{name of the response variable from \code{data} as a character string}

\item{Z.char}{names of covariates from \code{data} as vector of character strings}

\item{prov.char}{name of provider IDs variable as a character string}

\item{cutoff}{cutoff of provider size as an interger, default value is 10}
}
\description{
\code{fe.data.prep} prepares the data for model fitting with \code{fe.prov} by
taking the data with missing values imputed.
}
\seealso{
\code{\link{fe.data.prep}},  \code{\link{fe.prov}},   \code{\link{test.fe.prov}},
\code{\link{funnel.SRR}},   \code{\link{confint.fe.prov}}
}
