% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{Simulated readmissions data for 500 hospitals}
\format{A \code{data.frame} with 24438 rows and 5 variables (columns):
\describe{
\item{Y}{Indicator for readmission; 1=Yes, 0=No; numeric}
\item{prov.ID}{Provider ID; numeric}
\item{z1}{Simulated covariate 1, numeric}
\item{z2}{Simulated covariate 2, numeric}
\item{z3}{Simulated covariate 3, numeric}
}}
\usage{
hospital
}
\description{
A data set containing simulated readmissions data for 500 hospitals with
three continuous covariates. This data needs to be processed with \code{fe.data.prep}.
}
\keyword{datasets}
