% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.fe.prov.R
\name{confint.fe.prov}
\alias{confint.fe.prov}
\title{Compute confidence intervals for fitted model}
\usage{
\method{confint}{fe.prov}(data, fe.ls, Y.char, Z.char, prov.char, alpha)
}
\arguments{
\item{data}{prepared \code{data.frame}. Use \code{\link{fe.data.prep}}}

\item{fe.ls}{fitted model object (fit using \code{fe.prov})}

\item{Y.char}{Y.char name of the response variable from \code{data} as a character string}

\item{Z.char}{Z.char names of covariates from \code{data} as vector of character strings}

\item{prov.char}{name of provider IDs variable as a character string}

\item{alpha}{alpha level for the CIs}
}
\value{
Returns a \code{data.frame} of gamma and SRR lower and upper CI bounds. Each row is a
parameter, each column gives a different bound.
}
\description{
\code{confint.fe.prov} computes the (1-alpha)% confidence interval for the fixed
effect parameter estimates
}
\examples{
## data simulation
m <- 500
prov.size <- pmax(round(rnorm(m, 50, 15)),11)
gamma <- rnorm(m, log(3/7), 0.4)
beta <- c(1,0.5,-1)
Y.char <- 'Y'
prov.char <- 'prov.ID'
Z.char <- paste0('z', 1:length(beta))
data <- sim.fe.prov(m, prov.size, gamma, beta, Y.char, Z.char, prov.char)

library(Matrix)
library(poibin)
library(ggplot2)
cutoff <- 10              # an integer as cutoff of facility (or provider) size with 10 as default
tol <- 1e-5               # a small positive number specifying stopping criterion of Newton-Raphson algorithm
n <- 10000                # resample size, 10000 as default
alpha <- 0.05             # significance level
data.prepared <- fe.data.prep(data, Y.char, Z.char, prov.char, cutoff) # data preparation
fe.ls <- fe.prov(data.prepared, Y.char, Z.char, prov.char, tol) # model fitting

# confidence intervals
confint.fe.prov(data.prepared, fe.ls, Y.char, Z.char, prov.char, alpha)

}
\seealso{
\code{\link{fe.data.prep}},  \code{\link{fe.prov}},   \code{\link{test.fe.prov}},
\code{\link{funnel.SRR}},   \code{\link{confint.fe.prov}}
}
